// frontend/src/app/core/services/api/model/qualificationSchemeRequest.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Request payload for creating or updating a qualification scheme
 */
export interface QualificationSchemeRequest { 
    name: string;
    laneCount: number;
    isActive: boolean;
}